# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cash-Track API
  version: 1.0.0
  license:
    name: Proprietary
    url: https://cash-track.app
  description: |
    Cash-Track is a personal finance management SaaS. This API powers the mobile and web clients.

    ## Authentication
    Most endpoints require a **JWT Bearer token** obtained via `/v1/auth/login`, `/v1/auth/register`,
    or one of the social/passkey login flows.

    Tokens come in pairs:
    - **Access token** — short-lived (~1 hour). Send in `Authorization: Bearer <token>`.
    - **Refresh token** — long-lived (~7 days). Use `POST /v1/auth/refresh` to get a new pair.

    ## Email Confirmation
    After registration the user receives a confirmation email. Creating wallets, charges, and tags
    requires a confirmed email address. Use the `/v1/auth/email/confirmation` group to check status
    or resend the link.

    ## Error Responses
    Standard errors return an `Error` object. Validation failures return a `ValidationError`
    object with per-field messages (HTTP 422).

    ## Idempotency
    Most state-changing endpoints (`POST`/`PUT`/`PATCH`/`DELETE`) accept an optional
    `Idempotency-Key` request header — a client-generated UUIDv4. Retrying the exact same
    request with the same key within 24 hours replays the original response instead of
    re-executing it (`Idempotency-Replayed: true` is added to the replayed response). Reusing
...
    rejected (`400`). See `IdempotencyKeyParam`, `IdempotencyKeyConflict`, and
...

...

...

...
servers:
...
paths:
...
  /v1/wallets/{walletId}/charges:
    parameters:
      - $ref: "#/components/parameters/WalletIdQueryParam"

    get:
      operationId: chargeList
      summary: List wallet charges
      description: |
        Returns a paginated list of charges for the wallet, ordered newest first.
        Use `page` and `limit` for pagination.
      tags: [Charges]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Paginated charges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargePaginatedResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      operationId: chargeCreate
      summary: Create a charge
      description: |
        Records a new financial transaction. Requires a confirmed email.
        `type` is `+` for income and `-` for expense.
        `dateTime` defaults to the current server time when omitted and must be in the past.
      tags: [Charges]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChargeRequest"
            examples:
              expense:
                summary: Expense charge
                value:
                  type: "-"
                  amount: 42.50
                  title: Lunch
                  description: Sushi with the team
                  tags: [1, 3]
                  dateTime: "2024-06-01 13:00:00"
              income:
                summary: Income charge
                value:
                  type: "+"
                  amount: 3000.00
                  title: Salary
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Created charge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charge"
        "422":
          $ref: "#/components/responses/ValidationOrIdempotencyMismatch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
  /v1/wallets/{walletId}/charges/{chargeId}:
    parameters:
      - $ref: "#/components/parameters/WalletIdQueryParam"
      - $ref: "#/components/parameters/ChargeIdParam"

    put:
      operationId: chargeUpdate
      summary: Update a charge
      description: Replaces all mutable fields of a charge. All listed fields are required.
      tags: [Charges]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChargeRequest"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Updated charge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charge"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationOrIdempotencyMismatch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
    delete:
      operationId: chargeDelete
      summary: Delete a charge
      description: Permanently removes a charge. The wallet total is recalculated.
      tags: [Charges]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Charge deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
        "422":
          $ref: "#/components/responses/IdempotencyKeyMismatch"
  /v1/wallets/{walletId}/charges/move/{targetWalletId}:
    post:
      operationId: chargeMove
      summary: Move charges to another wallet
      description: |
        Moves one or more charges from this wallet to a target wallet.
        The caller must have access to both wallets.
      tags: [Charges]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WalletIdQueryParam"
        - name: targetWalletId
          in: path
          required: true
          description: Destination wallet ID
          schema:
            type: integer
        - $ref: "#/components/parameters/IdempotencyKeyParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chargeIds]
              properties:
                chargeIds:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    format: uuid
                  description: IDs of charges to move
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          description: Charges moved
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationOrIdempotencyMismatch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/IdempotencyKeyConflict"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/v1/auth/login`, `/v1/auth/register`, or any social
        login endpoint. Tokens expire after ~1 hour. Refresh via `POST /v1/auth/refresh`.

  # -------------------------------------------------------------------------
  # PARAMETERS
  # -------------------------------------------------------------------------
  parameters:
...
    WalletIdQueryParam:
      name: walletId
      in: path
      required: true
      description: Wallet integer ID
      schema:
        type: integer
        minimum: 1
    ChargeIdParam:
      name: chargeId
      in: path
      required: true
      description: Charge UUID
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional. A canonical lowercase UUIDv4 supplied by the client. Sending the same key
        again with the same method, path, and request body within 24 hours replays the first
        response verbatim (see the `Idempotency-Replayed` response header) instead of
        re-executing the request. Sending it again with a *different* body returns `422`
        instead of running — see the `IdempotencyKeyMismatch` response. A malformed value
        (not a canonical UUIDv4) returns `400`. Omitting the header entirely runs the request
        normally, with no deduplication.
      schema:
        type: string
        format: uuid
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
        example: "1e6c2a6a-9e2e-4d3a-8f2e-6a2f1b8c9d10"

  # -------------------------------------------------------------------------
  # SCHEMAS
  # -------------------------------------------------------------------------
  schemas:
...
    UserShort:
      type: object
      description: Minimal user representation used in shared contexts.
      required:
        - type
        - id
        - name
        - nickName
      properties:
        type:
          type: string
          enum: [userShort]
          readOnly: true
        id:
          type: integer
          readOnly: true
        name:
          type: string
        lastName:
          type: [string, "null"]
        nickName:
          type: string
        photoUrl:
          type: [string, "null"]
          format: uri
...
    WalletShort:
      type: object
      description: Minimal wallet representation used as a nested field inside charges and limits.
      required:
        - type
        - id
        - name
        - slug
        - totalAmount
        - isActive
        - isPublic
        - isArchived
        - createdAt
        - updatedAt
      properties:
        type:
          type: string
          enum: [walletShort]
          readOnly: true
        id:
          type: integer
          readOnly: true
        name:
          type: string
        slug:
          type: string
        totalAmount:
          type: number
          format: float
        isActive:
          type: boolean
        isPublic:
          type: boolean
        isArchived:
          type: boolean
        defaultCurrencyCode:
          type: [string, "null"]
        defaultCurrency:
          $ref: "#/components/schemas/Currency"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    Charge:
      type: object
      description: A single financial transaction.
      required:
        - type
        - id
        - operation
        - amount
        - title
        - userId
        - walletId
        - dateTime
        - createdAt
        - updatedAt
      properties:
        type:
          type: string
          enum: [charge]
          readOnly: true
        id:
          type: string
          format: uuid
          readOnly: true
          description: UUID primary key
        operation:
          type: string
          enum: ["+", "-"]
          description: "`+` = income, `-` = expense"
        amount:
          type: number
          format: float
          minimum: 0.01
        title:
          type: string
        description:
          type: [string, "null"]
        userId:
          type: integer
          description: ID of the user who created the charge
        walletId:
          type: integer
        dateTime:
          type: string
          format: date-time
          description: When the transaction occurred (in the past)
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        user:
          $ref: "#/components/schemas/UserShort"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        wallet:
          $ref: "#/components/schemas/WalletShort"
    CreateChargeRequest:
      type: object
      required:
        - type
        - amount
        - title
      properties:
        type:
          type: string
          enum: ["+", "-"]
          description: "`+` = income, `-` = expense"
        amount:
          type: number
          format: float
          minimum: 0.01
        title:
          type: string
          minLength: 1
        description:
          type: [string, "null"]
        tags:
          type: array
          items:
            type: integer
          description: Tag IDs to attach (must belong to the user or accessible shared wallet)
        dateTime:
          type: string
          description: "Transaction date in `Y-m-d H:i:s` format (e.g. `2024-06-01 13:00:00`). Defaults to now."
          example: "2024-06-01 13:00:00"
    ChargePaginatedResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Charge"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    Tag:
      type: object
      description: A label attached to charges for categorisation.
      required:
        - type
        - id
        - name
        - userId
        - createdAt
        - updatedAt
      properties:
        type:
          type: string
          enum: [tag]
          readOnly: true
        id:
          type: integer
          readOnly: true
        name:
          type: string
          minLength: 3
          maxLength: 255
          description: No spaces allowed; unique per user
        icon:
          type: [string, "null"]
          description: Emoji or short string (1–10 characters)
          example: "🍔"
        color:
          type: [string, "null"]
          pattern: "^#[0-9A-Fa-f]{6}$"
          description: Hex color code
          example: "#FF5733"
        userId:
          type: integer
          description: Owner user ID
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    Currency:
      type: object
      description: A supported currency with live exchange rate.
      required:
        - type
        - id
        - code
        - name
        - char
        - rate
        - updatedAt
      properties:
        type:
          type: string
          enum: [currency]
          readOnly: true
        id:
          type: string
          description: ISO 4217 code (same as `code`)
          example: USD
        code:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 currency code
          example: USD
        name:
          type: string
          description: Full currency name
          example: US Dollar
        char:
          type: string
          description: Currency symbol
          example: "$"
        rate:
          type: number
          format: float
          description: Exchange rate relative to the base currency (USD)
        updatedAt:
          type: string
          format: date-time
          description: When the exchange rate was last updated

    # -----------------------------------------------------------------------
    # PAGINATION
    # -----------------------------------------------------------------------
    Pagination:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
          description: Total number of items across all pages
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean

    # -----------------------------------------------------------------------
    # COMMON RESPONSES
    # -----------------------------------------------------------------------
...
    Error:
      type: object
      description: Standard error envelope.
      required:
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code (optional)
        message:
          type: string
          description: Human-readable error description
    ValidationError:
      type: object
      description: Field-level validation failures (HTTP 422).
      required:
        - errors
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Map of field names to arrays of error messages
          example:
            email:
              - "Must be a valid email address"
            password:
              - "Must be at least 6 characters"

  # -------------------------------------------------------------------------
  # REUSABLE HEADERS
  # -------------------------------------------------------------------------
  headers:
    ApiVersion:
      description: |
        Git tag of the running API image (e.g. `v1.4.2`). Present on every response; omitted
        entirely (rather than sent empty) when the image was built without a tag.
      schema:
        type: string
    ApiSha:
      description: |
        Git commit SHA of the running API image. Present on every response; omitted entirely
        (rather than sent empty) when the image was built without a commit SHA.
      schema:
        type: string
    TraceId:
      description: |
        The active OpenTelemetry trace ID. Present on every response; omitted entirely (rather
        than sent empty or all-zeros) when tracing is disabled (no valid active span context).
      schema:
        type: string
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
    IdempotencyReplayed:
      description: |
        `true` when this response is a verbatim replay of the first request that used this
        `Idempotency-Key`, rather than a fresh execution. Absent on a fresh response.
      schema:
        type: string
        enum: ["true"]

  # -------------------------------------------------------------------------
  # REUSABLE RESPONSES
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: |
        Bad request — invalid input or business rule violation. Also returned when an
        `Idempotency-Key` header is present but not a canonical lowercase UUIDv4.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: INVALID_CREDENTIALS
            message: The provided credentials are incorrect.
    Unauthorized:
      description: Authentication required or token expired
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Unauthenticated.
    NotFound:
      description: Resource not found
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Resource not found.
    IdempotencyKeyConflict:
      description: |
        A request with this `Idempotency-Key` (same scope, method, path, and request body) is
        still being processed. Retry after the given delay; do not send a different body under
        the same key while this is in flight.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: The original request with this Idempotency-Key is still processing. Please retry shortly.
    IdempotencyKeyMismatch:
      description: |
        This `Idempotency-Key` was already used with a different request body (method, path, or
        payload). Idempotency keys are single-use per distinct request — retry with a new key,
        or resend the exact original request body to get the stored result.

        Deliberately `Error`-shaped (a plain `message`, no `errors` map) rather than
        `ValidationError`-shaped: this is a conflict over key reuse, not a field-level
        validation failure, and callers should not try to render it as one.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: This Idempotency-Key was already used with a different request body.
...
    ValidationOrIdempotencyMismatch:
      description: |
        `422` on this endpoint has two distinct causes, both possible since it accepts an
        `Idempotency-Key`:
          - **Field-level validation failure** — `ValidationError`-shaped, carries an `errors`
            map keyed by field name.
          - **Idempotency key reused with a different body** — `Error`-shaped, a plain
            `message`, no `errors` key.

        Discriminate by the presence of `errors` in the body, not by anything else — both
        causes share this single `422` status code.
      headers:
        X-Ct-Api-Version:
          $ref: "#/components/headers/ApiVersion"
        X-Ct-Api-Sha:
          $ref: "#/components/headers/ApiSha"
        X-Ct-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ValidationError"
              - $ref: "#/components/schemas/Error"
...
security:
  - bearerAuth: []